refactor: rename group-calcs dir to groups
