+++ log-viewer/modules/datagrid/groups/GroupCalcs.ts
